Redirection to new signup from auth/ui changes
