+++ src/components/layout/Navigation/index.tsx
@@ -184,9 +184,9 @@ const Navigation = () => {
       >
         <ChatSocket />
         {userProfile.authLoading ? (
-          <div className="container flex min-h-20 items-center justify-between px-7 py-4 lg:py-5 " />
+          <div className="flex min-h-20 items-center justify-between px-7 py-4 lg:py-5 " />
         ) : (
-          <div className="container flex items-center justify-between px-7 py-4 lg:py-5">
+          <div className="flex items-center justify-between px-4 sm:px-7 py-4 lg:py-5">
             <Link href="/" className="max-lg:hidden">
               <Logo />
             </Link>
